requirements.txt
# -*- coding: utf-8 -*-

# CORE FastAPI stack
fastapi==0.110.0
uvicorn[standard]==0.34.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.19
python-multipart==0.0.6
pydantic==2.0.3
pydantic_core==2.3.0
starlette==0.36.3
jinja2==3.1.2

# HTTP & Auth
httpx==0.24.1
requests==2.31.0
email-validator==2.0.0.post2
pyjwt==2.8.0
rsa==4.9

# Utilidades
anyio==3.7.1
certifi==2024.7.4
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.1.3
greenlet==2.0.2
h11==0.14.0
httptools>=0.6.3
idna==3.4
jsonschema==4.17.3
jsonschema-specifications==2024.10.1
markupsafe==2.1.3
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
typing_extensions>=4.8.0
ujson==5.8.0
urllib3==2.0.4
watchfiles==0.20.0
websockets==11.0.3
zipp==3.16.2

# Windows-only
pywin32==310; platform_system=="Windows"
pywinpty==2.0.15; platform_system=="Windows"

# Otros
pipreqs==0.4.13
tzdata==2023.2
pytz==2023.3